+++ spec/kitchen_spec/directions/move_exercises_to_eoc/v3_spec.rb
@@ -56,6 +56,33 @@ RSpec.describe Kitchen::Directions::MoveExercisesToEOC::V3 do
     )
   end
 
+  let(:book_with_review_exercises_in_module_with_title_children) do
+    book_containing(html:
+      <<~HTML
+        <div data-type="chapter">
+          <div data-type="page">
+            <div data-type="document-title" id="page1TitleId"><em data-effect="italics">Italics</em> & not italics with <sub>subscript</sub> and <sup>superscript</sup> text page title</div>
+            <section id="sectionId1" class="CLASSNAME">
+              <div data-type="exercise" id="exercise_id1">
+                <div data-type="problem" id="problem_id1">
+                  <p>exercise content 1</p>
+                </div>
+                <div data-type="solution" id="solution_id1">
+                  <p>Solution content</p>
+                </div>
+              </div>
+            </section>
+          </div>
+          <div data-type="page">
+            <div data-type="document-title" id="page2TitleId">page 2 title</div>
+            <section id="sectionId2" class="CLASSNAME">
+            </section>
+          </div>
+        </div>
+      HTML
+    )
+  end
+
   context 'when append_to is not null' do
     it 'works' do
       described_class.new.bake(chapter: book_with_review_exercises.chapters.first, metadata_source: metadata_element, append_to: append_to, klass: 'CLASSNAME')
@@ -67,6 +94,13 @@ RSpec.describe Kitchen::Directions::MoveExercisesToEOC::V3 do
     it 'works' do
       described_class.new.bake(chapter: book_with_review_exercises.chapters.first, metadata_source: metadata_element, klass: 'CLASSNAME')
       expect(book_with_review_exercises.chapters.first).to match_snapshot_auto
+    end
+  end
+
+  context 'when exercises are in module with titles chilren' do
+    it 'keeps the modules titles children while creating a section title' do
+      described_class.new.bake(chapter: book_with_review_exercises_in_module_with_title_children.chapters.first, metadata_source: metadata_element, append_to: append_to, klass: 'CLASSNAME')
+      expect(append_to).to match_snapshot_auto
     end
   end
 end