+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Add `appendix` to `contemporary-math`
 * Add support for module title element children (e.g. italics, sup, sub) to be kept while baking
 
 ## [v1.16.0] - 2022-09-09